tried optimizing the currentUser object. Added annotations to User class for persistence API
