+++ src/main/java/API/AuthstudyApplication.java
@@ -1,4 +1,4 @@
-package com.authstudy.authstudy;
+package API;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
